Splitting of files
